postgres_to_es/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
elasticsearch==7.9.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
pydantic==1.7.3
python-dotenv==0.15.0
redis==3.5.3
requests==2.25.0
urllib3==1.26.2
backoff==1.10.0
dataclasses==0.6
psycopg2-binary==2.8.6